feat: Add README and initial webfonts folder
This commit adds the README explaining what this repository is for, as
well as the first folder of experimentation I engaged in. If I end up
using this repo more often than I anticipate, at a later point I might
have to change the information architecture - we'll see how it goes.