feat: Add day 1 of SVG Tutorial
This is the first day of 25 days of an SVG tutorial advent calendar.
Which I'm starting about a week late. Typical, but I've wanted to learn
about SVGs for a while.

Much of the context for the work will be documented in the subfolders'
README, but I am notably reinforcing the subfolder convention for the
information architecture while continuing to experiment with the
note-taking process, since this is all iterative.

The webfonts effort I just sat down and hacked at by my own initiative.
This has an actual structure to follow, which reduces the activation
effort involved in the maintenance of doing this work.