feat: Add code and notes from chapter 2 of the rust book
There wasn't anything too surprising in this chapter, though it did
introduce a few conventions I could comprehend but haven't quite yet
internalised. Some small things involving enums but the book frequently
introduces things and explains just enough context to move on.

It mentions that a new concept or idea will be given more prominence
during a later chapter, so I'm not trying to dogmatically overlearn the
thing. I actually did most of this work about two days ago, but hadn't
the time to write the notes and commit it.